# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for UniMatch Copilot AI system",
        "title": "UniMatch Copilot API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/universities": {
            "get": {
                "description": "Fetch paginated list of universities with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "universities"
                ],
                "summary": "List universities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by country",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "meta": {
                                            "$ref": "#/definitions/response.Meta"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new university record manually",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "universities"
                ],
                "summary": "Create university",
                "parameters": [
                    {
                        "description": "University Information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUniversityRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/universities/crawl-active": {
            "get": {
                "description": "Get the number of universities currently pending crawl results",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "universities"
                ],
                "summary": "Count active crawl jobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/universities/crawl-all": {
            "post": {
                "description": "Submits async crawl jobs for universities that haven't been crawled in 24h",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "universities"
                ],
                "summary": "Trigger crawl for all universities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CrawlAllResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/swagger.SwaggerError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.CrawlAllResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "triggered": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CreateUniversityRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "acceptance_rate": {
                    "type": "number"
                },
                "available_majors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "counselor_notes": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "gpa_expectation_normalized": {
                    "type": "number"
                },
                "group_tag": {
                    "type": "string"
                },
                "ielts_min": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "qs_rank": {
                    "type": "integer"
                },
                "sat_required": {
                    "type": "boolean"
                },
                "scholarship_available": {
                    "type": "boolean"
                },
                "scholarship_notes": {
                    "type": "string"
                },
                "tuition_usd_per_year": {
                    "type": "integer"
                }
            }
        },
...
        "response.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean"
                },
                "has_prev": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorInfo"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "swagger.SwaggerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "NOT_FOUND"
                },
                "details": {
                    "type": "string",
                    "example": "User with ID 123 not found"
                },
                "message": {
                    "type": "string",
                    "example": "Resource not found"
                }
            }
        }
    }
}